requirements.txt
black==22.6.0
click==8.1.3
flake8==5.0.4
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
tomli==2.0.1
typing_extensions==4.3.0
Werkzeug==2.2.2
flask-mysqldb==1.0.1
pyyaml==6.0
flask-sqlalchemy
pymysql